// src/app/smog/service/model/measurementsColorAndAvarage.ts
import { AverageAndColor } from './averageAndColor';
export interface MeasurementsColorAndAverage {
  airQualityIndex?: AverageAndColor;
  pm1?: AverageAndColor;
  pm25?: AverageAndColor;
  pm10?: AverageAndColor;
  pollutionLevel?: AverageAndColor;
  pressure?: AverageAndColor;
  humidity?: AverageAndColor;
  temperature?: AverageAndColor;
}